+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,8 +1,11 @@
-﻿### Version 3.8.11
+﻿### Version 3.8.12
+
+- Fixed: `NormalizeNamespace` has been fixed to ensure that it parses C# types correctly.
+
+### Version 3.8.11
 
 - Improvement: Added support for `Domain` and `Services` naming conventions for `Entities`, `Attributes` and `Operations`.
 - Fixed: `ObjectUpdateMapping` would crash in the event where `Children` mappings were `null`.
-- Fixed: `NormalizeNamespace` has been fixed to ensure that it parses C# types correctly.
 
 ### Version 3.8.10